deduplicate code to extrapolate points for bezier curves
