+++ client-data/board.html
@@ -85,6 +85,7 @@
 	<script src="../js/path-data-polyfill.js"></script>
 	<script src="../js/minitpl.js"></script>
 	<script src="../js/board.js"></script>
+	<script src="../tools/pencil/pencil_extrapolate_points.js"></script>
 	<script src="../tools/pencil/pencil.js"></script>
 	<script src="../tools/cursor/cursor.js"></script>
 	<script src="../tools/line/line.js"></script>
